requirements.txt
pytest~=8.3.3
allure-pytest==2.13.5
selene>=2.0.0rc9